Backend/PSM/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
mysqlclient>=2.2.0
pymongo>=4.6.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
whitenoise>=6.6.0
gunicorn>=21.2.0
